Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Not sure if relevant but I have seen issues with apps caused by some code being reloaded due to screen orientation changes. Something to keep in mind

Those have to be fixed properly though - this is a hack to not have to deal with the issue, which is not something we should do, as it can bite us.

Usually it's apps that don't like configuration changes - rotation is just one of those, there's others - language, theme, maybe more.


If there is an issue and we need to deploy this temporarily when fixing that fine, but otherwise we shouldn't be doing it this way.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>